added the installation script to be inside the utility
